refactor(settings): express-free service layer
