+++ api/src/settings/router.ts
@@ -1,355 +1,83 @@
-import crypto from 'crypto'
-import express from 'express'
-import nanoid from '../misc/utils/nanoid.js'
-import slug from 'slugify'
-import dayjs from 'dayjs'
-import equal from 'fast-deep-equal'
+import express, { type Request } from 'express'
 import { httpError } from '@data-fair/lib-utils/http-errors.js'
-import { type OptionsDesMetadonneesDeJeuxDeDonnees, type Settings } from '#types/settings/index.js'
-import { type DepartmentSettings } from '#types/department-settings/index.js'
 import * as cacheHeaders from '../misc/utils/cache-headers.js'
-import * as topicsUtils from '../misc/utils/topics.ts'
 import * as notifications from '../misc/utils/notifications.ts'
 import config from '#config'
-import mongo from '#mongo'
-import standardLicenses from '../../contract/licenses.js'
-import debugLib from 'debug'
 import { reqHost } from '@data-fair/lib-express/req-origin.js'
-import { type AccountKeys, reqSessionAuthenticated, reqUserAuthenticated, type SessionStateAuthenticated } from '@data-fair/lib-express'
-import { type Request } from '#types'
-import eventsLog from '@data-fair/lib-express/events-log.js'
-import eventsQueue from '@data-fair/lib-node/events-queue.js'
-import clone from '@data-fair/lib-utils/clone.js'
-import { validateSettings, cleanSettings, fillSettings, cleanDatasetsMetadata, isMainSettings, isDepartmentSettings } from './operations.ts'
+import { reqSessionAuthenticated } from '@data-fair/lib-express'
+import { reqEventLogContext } from '../misc/utils/req-context.ts'
+import { buildPublicationSiteSubscriptions } from './operations.ts'
+import * as service from './service.ts'
+import { type SettingsWriteContext } from './service.ts'
 import { settingsParamsMiddleware, isOwnerAdmin, isOwnerMember, reqSettingsParams } from './middlewares.ts'
 // TODO(task 1.4): remove this shim once api-key.ts imports operations.ts directly
 export { isMainSettings, isUserSettings, isDepartmentSettings } from './operations.ts'
-
-const debugPublicationSites = debugLib('publication-sites')
 
 const router = express.Router()
 
 router.use('/:type/:id', settingsParamsMiddleware)
 
-// read settings as owner
-router.get('/:type/:id', isOwnerAdmin, cacheHeaders.noCache, async (req, res) => {
-  const { ownerFilter } = reqSettingsParams(req)
-  const settings = mongo.settings
-  const result = await settings
-    .findOne(ownerFilter, { projection: { _id: 0, id: 0, type: 0 } })
-  res.status(200).send(result ? cleanSettings(result) : {})
+const reqWriteContext = (req: Request): SettingsWriteContext => ({
+  ...reqSettingsParams(req),
+  sessionState: reqSessionAuthenticated(req),
+  host: reqHost(req),
+  logCtx: reqEventLogContext(req)
 })
 
-const writeSettings = async (req: Request, existingSettings: Settings | DepartmentSettings | null, settings: any, sessionState: SessionStateAuthenticated) => {
-  const { owner, ownerFilter } = reqSettingsParams(req)
-  const user = sessionState.user
-  fillSettings(owner, user, settings)
-  validateSettings(settings)
-
-  settings.apiKeys = settings.apiKeys ?? []
-  const existingApiKeys = existingSettings?.apiKeys ?? []
-
-  // a copy of the api keys where the clearKey is returned for the user that created a new key
-  const returnedApiKeys = settings.apiKeys?.map(apiKey => ({ ...apiKey })) || []
-
-  for (let i = 0; i < settings.apiKeys.length; i++) {
-    const apiKey = settings.apiKeys[i]
-
-    // this check should not be necessary as later on we check adminMode at creation then immutability
-    // this is just here as an extra safety
-    if (apiKey.adminMode && !user.isAdmin) {
-      eventsLog.alert('df.apikeys.manageadmin', 'a non-admin user attempted to manage settings that include an adminMode api key', { req, account: owner })
-      throw httpError(403, 'Only superadmin can manage api keys with adminMode=true')
-    }
-
-    if (apiKey.key) {
-      eventsLog.alert('df.apikeys.writesecret', 'a user attempted to write an api key internal secret', { req, account: owner })
-      throw httpError(403, 'Attempt to write an api key secret')
-    }
-
-    if (apiKey.notifiedJ3At !== undefined || apiKey.notifiedJAt !== undefined) {
-      eventsLog.alert('df.apikeys.writeflag', 'a user attempted to write an api key internal notification flag', { req, account: owner })
-      throw httpError(400, 'API key notification flags are internal and not user-writable')
-    }
-
-    if (!apiKey.id) {
-      // creating a new key
-
-      const returnedApiKey = returnedApiKeys[i]
-      if (apiKey.adminMode && !user.adminMode) {
-        eventsLog.alert('df.apikeys.createadmin', 'a user attempted to create an adminMode api key', { req, account: owner })
-        throw httpError(403, 'Only superadmin can create api keys with adminMode=true')
-      }
-      if (apiKey.email) {
-        eventsLog.alert('df.apikeys.setemail', 'a user attempted to define the email address of an api key', { req, account: owner })
-        throw httpError(403, 'API key email is readonly')
-      }
-      if (apiKey.expireAt && apiKey.expireAt > dayjs().add(config.apiKeysMaxDuration + 1, 'day').format('YYYY-MM-DD')) {
-        throw httpError(400, 'API key expiration is too far in the future')
-      }
-      returnedApiKey.id = apiKey.id = nanoid()
-
-      const clearKeyParts = [owner.type.slice(0, 1), owner.id]
-      if (owner.department) clearKeyParts.push(owner.department)
-      clearKeyParts.push(nanoid())
-      returnedApiKey.clearKey = Buffer.from(clearKeyParts.join(':')).toString('base64url')
-      const hash = crypto.createHash('sha512')
-      hash.update(returnedApiKey.clearKey)
-      returnedApiKey.key = apiKey.key = hash.digest('hex')
-
-      if (settings.type === 'user' && apiKey.scopes.length) {
-        returnedApiKey.email = apiKey.email = (settings as Settings).email
-      } else {
-        returnedApiKey.email = apiKey.email = `${slug.default(apiKey.title, { lower: true, strict: true })}-${apiKey.id}@api-key.${reqHost(req)}`
-      }
-
-      eventsLog.info('df.apikeys.create', `a user created an api key ${apiKey.title} (${apiKey.id}), scopes=${apiKey.scopes.join(', ')}`, { req, account: owner })
-      eventsQueue.pushEvent({
-        title: 'Création d\'une clé d\'API',
-        body: `${apiKey.title} (${apiKey.id}), scopes=${apiKey.scopes.join(', ')}`,
-        topic: {
-          key: 'data-fair:settings:api-key-created'
-        },
-        sender: owner
-      }, sessionState)
-    } else {
-      // re-sending an existing key
-
-      const existingApiKey = existingApiKeys.find(k => k.id === apiKey.id)
-      if (!existingApiKey) {
-        eventsLog.alert('df.apikeys.setid', 'a user tried to create an api key with id', { req, account: owner })
-        throw httpError(400, 'API key cannot be created with an id')
-      }
-      // should be covered by next general immutability check, but double check to be sure
-      if (apiKey.adminMode && !existingApiKey.adminMode && !user.adminMode) {
-        eventsLog.alert('df.apikeys.setadmin', 'a user attempted to mutate an api key and make it admin', { req, account: owner })
-        throw httpError(403, 'Only superadmin can delete api keys with adminMode=true')
-      }
-      if (existingApiKey.key) {
-        apiKey.key = existingApiKey.key
-      } else {
-        eventsLog.warn('df.apikeys.missingKey', `an API ${apiKey.id} key seems to be missing its internal secret`, { req, account: owner })
-      }
-      // notifiedJ3At / notifiedJAt are internal flags set by the expiration worker; they are
-      // stripped from API responses and never sent back by the user, so exclude them from the
-      // immutability comparison (otherwise any key the worker has notified becomes un-resavable).
-      const { notifiedJ3At: _nj3, notifiedJAt: _nj, ...comparableExistingApiKey } = existingApiKey as any
-      if (!equal(comparableExistingApiKey, apiKey)) {
-        eventsLog.alert('df.apikeys.mutate', `a user tried to mutate an existing api key ${existingApiKey.title} (${existingApiKey.id})`, { req, account: owner })
-        throw httpError(400, 'existing API keys are immutable')
-      }
-    }
-  }
-
-  // deleting an api key
-  for (const existingApiKey of existingApiKeys) {
-    if (!settings.apiKeys.some(k => k.id === existingApiKey.id)) {
-      if (existingApiKey.adminMode && !user.adminMode) {
-        eventsLog.alert('df.apikeys.deleteadmin', 'a user attempted to delete an admin api key', { req, account: owner })
-        throw httpError(403, 'Only superadmin can delete api keys with adminMode=true')
-      }
-      eventsQueue.pushEvent({
-        title: 'Suppression d\'une clé d\'API',
-        body: `${existingApiKey.title} (${existingApiKey.id}), scopes=${existingApiKey.scopes.join(', ')}`,
-        topic: {
-          key: 'data-fair:settings:api-key-deleted'
-        },
-        sender: owner
-      }, sessionState)
-    }
-  }
-
-  if (isMainSettings(settings) && settings.privateVocabulary) {
-    for (const concept of settings.privateVocabulary) {
-      if (!concept.id) concept.id = slug.default(concept.title, { lower: true, strict: true })
-      if (!concept.identifiers || !concept.identifiers.length) concept.identifiers = [concept.id]
-    }
-  }
-
-  if (isMainSettings(settings) && settings.topics) {
-    for (const topic of settings.topics) {
-      if (!topic.id) topic.id = nanoid()
-    }
-  }
-
-  if (isMainSettings(settings) && settings.datasetsMetadata) {
-    cleanDatasetsMetadata(settings.datasetsMetadata)
-  }
-  const oldSettings = (await mongo.settings.findOneAndReplace(ownerFilter, settings, { upsert: true }))
-
-  if (oldSettings && isMainSettings(oldSettings) && isMainSettings(settings) && settings.topics) {
-    await topicsUtils.updateTopics(owner, oldSettings.topics || [], settings.topics)
-  }
-
-  if (oldSettings && isMainSettings(oldSettings) && isMainSettings(settings) && settings.datasetsMetadata) {
-    await updateDatasetsMetadata(owner, oldSettings.datasetsMetadata || {}, settings.datasetsMetadata)
-  }
-
-  return cleanSettings({ ...settings, apiKeys: returnedApiKeys })
-}
-
-const updateDatasetsMetadata = async (owner: AccountKeys, oldDatasetsMetadata: OptionsDesMetadonneesDeJeuxDeDonnees, newDatasetsMetadata: OptionsDesMetadonneesDeJeuxDeDonnees) => {
-  if (equal(oldDatasetsMetadata, newDatasetsMetadata)) return
-  for (const oldMeta of oldDatasetsMetadata.custom ?? []) {
-    if (newDatasetsMetadata.custom?.some(nc => nc.key === oldMeta.key)) {
-      await mongo.datasets.updateMany(
-        { 'owner.type': owner.type, 'owner.id': owner.id, [`customMetadata.${oldMeta.key}`]: { $exists: true } },
-        { $unset: { [`customMetadata.${oldMeta}`]: 1 } })
-      await mongo.datasets.updateMany(
-        { 'owner.type': owner.type, 'owner.id': owner.id, [`draft.customMetadata.${oldMeta.key}`]: { $exists: true } },
-        { $unset: { [`draft.customMetadata.${oldMeta}`]: 1 } })
-    }
-  }
-}
+// read settings as owner
+router.get('/:type/:id', isOwnerAdmin, cacheHeaders.noCache, async (req, res) => {
+  res.status(200).send(await service.getSettings(reqSettingsParams(req).ownerFilter))
+})
 
 // update settings as owner
 router.put('/:type/:id', isOwnerAdmin, async (req, res) => {
-  const { ownerFilter } = reqSettingsParams(req)
-  const settings = req.body
-  const sessionState = reqSessionAuthenticated(req)
-  const existingSettings = await mongo.settings.findOne(ownerFilter)
-  const returnedSettings = await writeSettings(req, existingSettings, settings, sessionState)
-  res.status(200).send(returnedSettings)
+  res.status(200).send(await service.updateSettings(reqWriteContext(req), req.body))
 })
 router.patch('/:type/:id', isOwnerAdmin, async (req, res) => {
-  const { ownerFilter } = reqSettingsParams(req)
-  const partialSettings = req.body
-  const sessionState = reqSessionAuthenticated(req)
-  const existingSettings = await mongo.settings.findOne(ownerFilter, { projection: { _id: 0 } })
-  const settings = cleanSettings({ ...(clone(existingSettings ?? {})), ...partialSettings })
-  const returnedSettings = await writeSettings(req, existingSettings, settings, sessionState)
-  res.status(200).send(returnedSettings)
+  res.status(200).send(await service.patchSettings(reqWriteContext(req), req.body))
 })
 
 // Get topics list as owner
 router.get('/:type/:id/topics', isOwnerMember, async (req, res) => {
-  const { ownerFilter } = reqSettingsParams(req)
-  const settings = mongo.settings
-  const result = await settings.findOne(ownerFilter)
-  const topics = result && isMainSettings(result) && result.topics ? result.topics : []
-  for (const topic of topics) {
-    if (topic.icon) topic.icon = { name: topic.icon.name, svgPath: topic.icon.svgPath }
-  }
-  res.status(200).send(topics)
+  res.status(200).send(await service.getTopics(reqSettingsParams(req).ownerFilter))
 })
 
 // Get licenses list as anyone
 router.get('/:type/:id/licenses', cacheHeaders.noCache, async (req, res) => {
-  const { ownerFilter } = reqSettingsParams(req)
-  const settings = mongo.settings
-  const result = await settings.findOne(ownerFilter)
-  const licenses: { href: string, title: string }[] = []
-  for (const l of standardLicenses) {
-    licenses.push({ href: l.href, title: l.title })
-  }
-  if (result && isMainSettings(result)) {
-    for (const l of result.licenses ?? []) {
-      licenses.push(l)
-    }
-  }
-
-  res.status(200).send(licenses)
+  res.status(200).send(await service.getLicenses(reqSettingsParams(req).ownerFilter))
 })
 
 // Get datasets metadata settings as owner
 router.get('/:type/:id/datasets-metadata', isOwnerMember, async (req, res) => {
-  const { ownerFilter } = reqSettingsParams(req)
-  const settings = mongo.settings
-  const result = await settings.findOne(ownerFilter)
-  res.status(200).send(result && isMainSettings(result) && result.datasetsMetadata ? result.datasetsMetadata : {})
+  res.status(200).send(await service.getDatasetsMetadata(reqSettingsParams(req).ownerFilter))
 })
 
 // Get agent chat setting as member
 router.get('/:type/:id/agent-chat', isOwnerMember, cacheHeaders.noCache, async (req, res) => {
-  const { ownerFilter } = reqSettingsParams(req)
-  const result = await mongo.settings.findOne(ownerFilter, { projection: { _id: 0, agentChat: 1 } })
-  res.status(200).send({ agentChat: !!(result && isMainSettings(result) && result.agentChat) })
+  res.status(200).send(await service.getAgentChat(reqSettingsParams(req).ownerFilter))
 })
 
 // Get publication sites as owner
 router.get('/:type/:id/publication-sites', isOwnerMember, async (req, res) => {
-  const { owner, ownerFilter, department } = reqSettingsParams(req)
   reqSessionAuthenticated(req)
-  const filter = [ownerFilter]
-  if (owner.department) {
-    filter.push({ ...ownerFilter, department: { $exists: false } })
-  } else if (department === '*') {
-    filter[0] = { ...ownerFilter }
-    delete filter[0].department
-  }
-  const settingsArray = await mongo.settings.find({ $or: filter }, { projection: { _id: 0 } }).toArray()
-  const publicationSites = []
-  for (const settings of settingsArray) {
-    for (const publicationSite of settings.publicationSites || []) {
-      if (isDepartmentSettings(settings)) publicationSite.department = settings.department
-      if (owner.department && publicationSite.settings?.contributorDepartments?.includes(owner.department)) {
-        (publicationSite as any).canContributeAsDepartment = true
-      }
-      publicationSites.push(publicationSite)
-    }
-  }
-  res.status(200).send(publicationSites)
+  res.status(200).send(await service.getPublicationSites(reqSettingsParams(req)))
 })
 // create/update publication sites as owner (used by data-fair-portals to sync portals)
 router.post('/:type/:id/publication-sites', isOwnerAdmin, async (req, res) => {
-  const { owner, ownerFilter } = reqSettingsParams(req)
-  debugPublicationSites('post site', req.body)
-  if (owner.department && req.body.settings?.contributorDepartments && req.body.settings.contributorDepartments.length) {
-    throw httpError(400, 'contributorDepartments is only allowed on org-root publication sites')
-  }
-  let settings = (await mongo.settings.findOne(ownerFilter, { projection: { _id: 0 } })) as Settings | DepartmentSettings
-  if (!settings) {
-    settings = fillSettings(owner, reqUserAuthenticated(req), {})
-  }
-  settings.publicationSites = settings.publicationSites || []
-  const index = settings.publicationSites.findIndex(ps => ps.type === req.body.type && ps.id === req.body.id)
-  if (index === -1) {
-    settings.publicationSites.push(req.body)
-    const baseSubscription = {
-      outputs: ['devices', 'email'],
-      locale: 'fr',
-      sender: owner,
-      visibility: 'private'
+  const ctx = reqWriteContext(req)
+  const { created } = await service.upsertPublicationSite(ctx, req.body)
+  if (created) {
+    for (const subscription of buildPublicationSiteSubscriptions(ctx.owner, req.body, config.publicUrl)) {
+      notifications.subscribe(req, subscription)
     }
-    notifications.subscribe(req, {
-      ...baseSubscription,
-      topic: {
-        key: `data-fair:dataset-publication-requested:${req.body.type}:${req.body.id}`,
-        title: `Un contributeur demande de publier un jeu de données sur ${req.body.title || req.body.url || req.body.id}`
-      },
-      urlTemplate: config.publicUrl + '/dataset/{id}'
-    })
-    notifications.subscribe(req, {
-      ...baseSubscription,
-      topic: {
-        key: `data-fair:application-publication-requested:${req.body.type}:${req.body.id}`,
-        title: `Un contributeur demande de publier une application sur ${req.body.title || req.body.url || req.body.id}`
-      },
-      urlTemplate: config.publicUrl + '/application/{id}'
-    })
-  } else {
-    settings.publicationSites[index] = { ...req.body, settings: { ...(settings.publicationSites[index].settings || {}), ...(req.body.settings || {}) } }
   }
-  validateSettings(settings)
-  await mongo.settings.replaceOne(owner, settings, { upsert: true })
   res.status(200).send(req.body)
 })
 // delete publication sites as owner (used by data-fair-portals to sync portals)
 router.delete('/:type/:id/publication-sites/:siteType/:siteId', isOwnerAdmin, async (req, res) => {
-  debugPublicationSites('delete site', req.params)
-  const { owner, ownerFilter } = reqSettingsParams(req)
-
-  let settings = (await mongo.settings.findOne(ownerFilter, { projection: { _id: 0 } })) as Settings | DepartmentSettings
-  if (!settings) {
-    settings = fillSettings(owner, reqUserAuthenticated(req), {})
-  }
-  settings.publicationSites = settings.publicationSites || []
-  settings.publicationSites = settings.publicationSites.filter(ps => ps.type !== req.params.siteType || ps.id !== req.params.siteId)
-  validateSettings(settings)
-  await mongo.settings.replaceOne(ownerFilter, settings, { upsert: true })
-  const ref = `${req.params.siteType}:${req.params.siteId}`
-  await mongo.datasets.updateMany({ publicationSites: ref }, { $pull: { publicationSites: ref } })
-  await mongo.applications.updateMany({ publicationSites: ref }, { $pull: { publicationSites: ref } })
+  const { siteType, siteId } = req.params
+  if (typeof siteType !== 'string' || typeof siteId !== 'string') throw httpError(400, 'invalid path parameters')
+  await service.deletePublicationSite(reqWriteContext(req), siteType, siteId)
   res.status(200).send(req.body)
 })
