+++ dev/type-errors-baseline.txt
@@ -1,1 +1,1 @@
-1803
+1800
